Move FSDP CUDA graph test
